# api/gen/go/admin/service/v1/i_identity_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.8.4
// - protoc             (unknown)
// source: admin/service/v1/i_identity.proto

package adminpb

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "github.com/hypercoze/kratos-admin/api/gen/go/identity/service/v1"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationIdentityCreateAdmin = "/admin.service.v1.Identity/CreateAdmin"
const OperationIdentityGetAdmin = "/admin.service.v1.Identity/GetAdmin"
const OperationIdentityListAdmin = "/admin.service.v1.Identity/ListAdmin"
const OperationIdentityUpdateAdmin = "/admin.service.v1.Identity/UpdateAdmin"

type IdentityHTTPServer interface {
	CreateAdmin(context.Context, *v1.CreateAdminRequest) (*v1.Admin, error)
	GetAdmin(context.Context, *v1.GetAdminRequest) (*v1.Admin, error)
	ListAdmin(context.Context, *v1.ListAdminRequest) (*v1.ListAdminResponse, error)
	UpdateAdmin(context.Context, *v1.UpdateAdminRequest) (*v1.Admin, error)
}

func RegisterIdentityHTTPServer(s *http.Server, srv IdentityHTTPServer) {
	r := s.Route("/")
	r.GET("/admin/v1/admin", _Identity_ListAdmin0_HTTP_Handler(srv))
	r.POST("/admin/v1/admin", _Identity_CreateAdmin0_HTTP_Handler(srv))
	r.PUT("/admin/v1/admin/{id}", _Identity_UpdateAdmin0_HTTP_Handler(srv))
	r.GET("/admin/v1/admin/{id}", _Identity_GetAdmin0_HTTP_Handler(srv))
}

func _Identity_ListAdmin0_HTTP_Handler(srv IdentityHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.ListAdminRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationIdentityListAdmin)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListAdmin(ctx, req.(*v1.ListAdminRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v1.ListAdminResponse)
		return ctx.Result(200, reply)
	}
}

func _Identity_CreateAdmin0_HTTP_Handler(srv IdentityHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.CreateAdminRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationIdentityCreateAdmin)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateAdmin(ctx, req.(*v1.CreateAdminRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v1.Admin)
		return ctx.Result(200, reply)
	}
}

func _Identity_UpdateAdmin0_HTTP_Handler(srv IdentityHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.UpdateAdminRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationIdentityUpdateAdmin)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateAdmin(ctx, req.(*v1.UpdateAdminRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v1.Admin)
		return ctx.Result(200, reply)
	}
}

func _Identity_GetAdmin0_HTTP_Handler(srv IdentityHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.GetAdminRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationIdentityGetAdmin)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetAdmin(ctx, req.(*v1.GetAdminRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v1.Admin)
		return ctx.Result(200, reply)
	}
}

type IdentityHTTPClient interface {
	CreateAdmin(ctx context.Context, req *v1.CreateAdminRequest, opts ...http.CallOption) (rsp *v1.Admin, err error)
	GetAdmin(ctx context.Context, req *v1.GetAdminRequest, opts ...http.CallOption) (rsp *v1.Admin, err error)
	ListAdmin(ctx context.Context, req *v1.ListAdminRequest, opts ...http.CallOption) (rsp *v1.ListAdminResponse, err error)
	UpdateAdmin(ctx context.Context, req *v1.UpdateAdminRequest, opts ...http.CallOption) (rsp *v1.Admin, err error)
}

type IdentityHTTPClientImpl struct {
	cc *http.Client
}

func NewIdentityHTTPClient(client *http.Client) IdentityHTTPClient {
	return &IdentityHTTPClientImpl{client}
}

func (c *IdentityHTTPClientImpl) CreateAdmin(ctx context.Context, in *v1.CreateAdminRequest, opts ...http.CallOption) (*v1.Admin, error) {
	var out v1.Admin
	pattern := "/admin/v1/admin"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationIdentityCreateAdmin))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *IdentityHTTPClientImpl) GetAdmin(ctx context.Context, in *v1.GetAdminRequest, opts ...http.CallOption) (*v1.Admin, error) {
	var out v1.Admin
	pattern := "/admin/v1/admin/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationIdentityGetAdmin))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *IdentityHTTPClientImpl) ListAdmin(ctx context.Context, in *v1.ListAdminRequest, opts ...http.CallOption) (*v1.ListAdminResponse, error) {
	var out v1.ListAdminResponse
	pattern := "/admin/v1/admin"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationIdentityListAdmin))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *IdentityHTTPClientImpl) UpdateAdmin(ctx context.Context, in *v1.UpdateAdminRequest, opts ...http.CallOption) (*v1.Admin, error) {
	var out v1.Admin
	pattern := "/admin/v1/admin/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationIdentityUpdateAdmin))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}