# api/gen/openapi/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /admin/v1/admin:
        get:
            tags:
                - Identity
                - Authentication
            summary: 后台用户-列表
            description: 后台用户列表
            operationId: Identity_ListAdmin
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: username
                  in: query
                  schema:
                    type: string
                - name: phone
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListAdminResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        post:
            tags:
                - Identity
                - Authentication
            summary: 后台用户-创建
            description: 创建后台用户
            operationId: Identity_CreateAdmin
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateAdminRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Admin'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /admin/v1/admin/{id}:
        get:
            tags:
                - Identity
                - Authentication
            summary: 后台用户-通过id获取
            description: 通过id获取后台用户
            operationId: Identity_GetAdmin
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Admin'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        put:
            tags:
                - Identity
                - Authentication
            summary: 后台用户-修改
            description: 修改后台用户
            operationId: Identity_UpdateAdmin
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateAdminRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Admin'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
        Admin:
            type: object
            properties:
                id:
                    type: string
                    description: 用户ID
                realName:
                    type: string
                    description: 真实姓名
                username:
                    type: string
                    description: 用户名
                email:
                    type: string
                    description: 邮箱
                phone:
                    type: string
                    description: 手机号
                avatar:
                    type: string
                    description: 头像
                status:
                    type: string
                    description: 状态(enabled/disabled)
                createTime:
                    type: string
                    description: 创建时间
                    format: date-time
                updateTime:
                    type: string
                    description: 更新时间
                    format: date-time
            description: 后台用户信息
        CreateAdminRequest:
            required:
                - username
                - password
            type: object
            properties:
                realName:
                    type: string
                    description: 真实姓名
                username:
                    type: string
                    description: 用户名
                email:
                    type: string
                    description: 邮箱
                phone:
                    type: string
                    description: 手机号
                avatar:
                    type: string
                    description: 头像
                password:
                    type: string
                    description: 密码
            description: 创建后台用户请求
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        ListAdminResponse:
            type: object
            properties:
                total:
                    type: string
                    description: 总数
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Admin'
                    description: 列表
            description: 后台用户分页结果
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        UpdateAdminRequest:
            required:
                - id
            type: object
            properties:
                id:
                    type: string
                    description: 用户ID
                realName:
                    type: string
                    description: 真实姓名
                email:
                    type: string
                    description: 邮箱
                phone:
                    type: string
                    description: 手机号
                avatar:
                    type: string
                    description: 头像
                status:
                    type: string
                    description: 状态(enabled/disabled)
            description: 更新后台用户请求
...